Finalize v2.0: Replace v1 implementation with composition architecture
Cleanup changes to complete the transition to v2.0:

File Structure Changes:
- Removed src/engine.v1.ts (old monolithic implementation)
- Renamed src/engine-v2.ts → src/engine.ts (v2 is now primary)
- Renamed src/engine-v2.spec.ts → src/engine.spec.ts
- Removed v1-specific test files (bugfixes.v1.spec.ts, enhanced-engine.v1.spec.ts)
- Updated src/index.ts to import from './engine' instead of './engine-v2'

Test Updates:
- Fixed TypeScript strict mode type annotations in test callbacks
- Changed Entity type references to EntityDef interface (proper abstraction)
- All 34 comprehensive tests passing

The codebase now exclusively uses the composition-based architecture.
Previous v1 implementation and tests have been removed from the active
codebase (still available in git history if needed for reference).

This completes the v2.0 architecture migration.
